demo resources: move static resources to the demo folder
