+++ src/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>BPMN Visualization JS</title>
-    <link href="./static/css/main.css" rel="stylesheet">
+    <link href="demo/static/css/main.css" rel="stylesheet">
 </head>
 <body>
     <div id="file-selector" class="hidden">
